Move rectagle into own class + fix dimension missing in rectangle range queries
